+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 - [fix:fix watch tsf config, fix bean refresh with RefreshScope and ConfigurationProperties.](https://github.com/Tencent/spring-cloud-tencent/pull/1511)
 - [docs:simplify GitHub Actions.](https://github.com/Tencent/spring-cloud-tencent/pull/1514)
 - [feat: support config event.](https://github.com/Tencent/spring-cloud-tencent/pull/1532)
+- [feat:support stat and event report with service discovery.](https://github.com/Tencent/spring-cloud-tencent/pull/1533)